+++ src/containers/searchPage/SearchBody.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react'
 
 import Card from '../../components/card/Card'
 import CardIcon from '../../components/listPage/cardIcon/CardIcon'
-import { DataType } from '../listPage/CardList'
+import { DataType } from '../listPage/cardList/CardList'
 import SearchCateDrop from '../../components/searchPage/searchCateDrop/SearchCateDrop'
 import SearchNav from '../../components/searchPage/searchNav/SearchNav'
 import CardBadge from '../../components/listPage/cardBadge/CardBadge'
